pyproject.toml
[tool.poetry]
name = "tabled-pdf"
version = "0.1.0"
description = "Detect and recognize tables in PDFs and images."
authors = ["Vik Paruchuri <github@vikas.sh>"]
readme = "README.md"
license = "GPL-3.0-or-later"
repository = "https://github.com/VikParuchuri/tabled"
keywords = ["table", "table-recognition", "ocr", "pdf"]
packages = [
    {include = "tabled"}
]
include = [
    "extract.py",
    "table_app.py",
    "run_table_app.py",
]

[tool.poetry.dependencies]
python = "^3.10"
surya-ocr = "^0.6.1"
click = "^8.1.7"
pypdfium2 = "^4.30.0"
pydantic-settings = "^2.5.2"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
datasets = "^3.0.1"
streamlit = "^1.39.0"

[tool.poetry.scripts]
tabled_gui = "run_table_app:run_app"
tabled = "extract:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
